Backend/requirements.txt
absl-py==2.1.0
ag2==0.7.2
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncer==0.0.8
attrs==24.2.0
Authlib==1.4.0
autogen==0.7.2
autogen-agentchat==0.4.3
autogen-core==0.4.3
awscli==1.36.31
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.35.90
botocore==1.35.90
build==1.2.2.post1
cachetools==5.5.0
camelot-py==0.9.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.18
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.3
curl_cffi==0.7.3
cvzone==1.6.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.8
decorator==5.1.1
Deprecated==1.2.15
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docutils==0.16
duckduckgo_search==5.1.0
durationpy==0.9
effdet==0.4.1
emoji==2.14.0
et_xmlfile==2.0.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
fast-depends==2.4.12
fastapi==0.115.5
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
genai==2.1.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-cloud-vision==3.8.1
google-crc32c==1.6.0
google-genai==0.2.2
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jax==0.4.37
jaxlib==0.4.36
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.3.7
langchain-chroma==0.2.0
langchain-community==0.3.7
langchain-core==0.3.19
langchain-google-genai==2.0.4
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.143
layoutparser==0.3.4
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.20
ml_dtypes==0.5.0
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.14.2
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.0
openai==1.60.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-asgi==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
opt_einsum==3.4.0
orjson==3.10.11
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pi_heif==0.20.0
pikepdf==9.4.2
pillow==11.0.0
platformdirs==4.3.6
playsound==1.3.0
portalocker==2.10.1
posthog==3.7.0
primp==0.8.1
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pyautogen==0.7.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.6.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.24.13
pypandoc==1.14
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.17
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.7.2
s3transfer==0.10.4
safetensors==0.4.5
scipy==1.14.1
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SpeechRecognition==3.12.0
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.41.2
streamlit==1.40.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.3.3
timm==1.0.11
tokenizers==0.20.3
toml==0.10.2
tomli==2.1.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.5
unstructured-client==0.27.0
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
vertexai==1.71.1
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.17.2
yt-dlp==2024.11.18
zipp==3.21.0
